share MockTarget + RevertingTarget
