+++ tests/live/test_ccip_composer_fork.py
@@ -34,7 +34,14 @@ from web3.exceptions import ContractLogicError, Web3RPCError
 
 from pydefi.types import Address
 from pydefi.vm import Program
-from tests.live.sol_utils import compile_sol_file, deploy, ensure_solc
+from tests.live.sol_utils import (
+    MOCK_REVERTING_TARGET_SOL,
+    MOCK_TARGET_SOL,
+    compile_sol_file,
+    compile_sol_source,
+    deploy,
+    ensure_solc,
+)
 
 # Reusable "this should revert" matcher for both anvil-direct and forked reverts.
 _REVERT = (ContractLogicError, Web3RPCError)
@@ -151,34 +158,6 @@ contract MockERC20 {
         return true;
     }
 }
-
-/// @notice Mock target contract — records the most recent call.
-contract MockTarget {
-    event Called(address sender, uint256 value, bytes data);
-
-    uint256 public callCount;
-    bytes public lastData;
-    uint256 public lastValue;
-
-    function execute(bytes calldata data) external payable returns (bool) {
-        callCount++;
-        lastData = data;
-        lastValue = msg.value;
-        emit Called(msg.sender, msg.value, data);
-        return true;
-    }
-
-    receive() external payable {}
-}
-
-/// @notice Mock target that always reverts.
-contract RevertingTarget {
-    error AlwaysReverts();
-
-    fallback() external payable {
-        revert AlwaysReverts();
-    }
-}
 """
 
 
@@ -192,8 +171,8 @@ def _compile_mock_contracts() -> dict[str, dict]:
     return {
         "MockRouter": result["<stdin>:MockRouter"],
         "MockERC20": result["<stdin>:MockERC20"],
-        "MockTarget": result["<stdin>:MockTarget"],
-        "RevertingTarget": result["<stdin>:RevertingTarget"],
+        "MockTarget": compile_sol_source(MOCK_TARGET_SOL, "MockTarget"),
+        "RevertingTarget": compile_sol_source(MOCK_REVERTING_TARGET_SOL, "RevertingTarget"),
     }
