+++ tests/live/test_oft_composer_fork.py
@@ -34,7 +34,15 @@ from web3.exceptions import ContractLogicError, Web3RPCError
 
 from pydefi.types import Address
 from pydefi.vm import Program
-from tests.live.sol_utils import compile_sol_file, deploy, deploy_mock_v3_pool, ensure_solc
+from tests.live.sol_utils import (
+    MOCK_REVERTING_TARGET_SOL,
+    MOCK_TARGET_SOL,
+    compile_sol_file,
+    compile_sol_source,
+    deploy,
+    deploy_mock_v3_pool,
+    ensure_solc,
+)
 
 
 def _start_program() -> tuple[Program, "object", "object"]:
@@ -184,34 +192,6 @@ contract MockOFTAdapter {
         return _token;
     }
 }
-
-/// @notice Mock target contract — records the most recent call and emits an event.
-contract MockTarget {
-    event Called(address sender, uint256 value, bytes data);
-
-    uint256 public callCount;
-    bytes public lastData;
-    uint256 public lastValue;
-
-    function execute(bytes calldata data) external payable returns (bool) {
-        callCount++;
-        lastData = data;
-        lastValue = msg.value;
-        emit Called(msg.sender, msg.value, data);
-        return true;
-    }
-
-    receive() external payable {}
-}
-
-/// @notice Mock target that always reverts — used to test sub-call failure handling.
-contract RevertingTarget {
-    error AlwaysReverts();
-
-    fallback() external payable {
-        revert AlwaysReverts();
-    }
-}
 """
 
 
@@ -227,8 +207,8 @@ def _compile_mock_contracts() -> dict[str, dict]:
         "MockEndpoint": result["<stdin>:MockEndpoint"],
         "MockOFT": result["<stdin>:MockOFT"],
         "MockOFTAdapter": result["<stdin>:MockOFTAdapter"],
-        "MockTarget": result["<stdin>:MockTarget"],
-        "RevertingTarget": result["<stdin>:RevertingTarget"],
+        "MockTarget": compile_sol_source(MOCK_TARGET_SOL, "MockTarget"),
+        "RevertingTarget": compile_sol_source(MOCK_REVERTING_TARGET_SOL, "RevertingTarget"),
     }
